Set backendHostHeader to AFD hostname from Frontdoor config
Backend host header should not be set unless necessary, if left blank it
will just pass through the original host header from the request, which
is usually the wanted behaviour.
https://docs.microsoft.com/de-de/azure/frontdoor/front-door-backend-pool#hostheader

AFD does health-checks with the backends hostname, which breaks some
backend configurations. Here we need to explicity set the AFD name as a
backendHostHeader.